Remove PreDeployInstance class
